chore: Move tokio runtime state to state file
